+++ lib/chronic.rb
@@ -1,24 +1,30 @@
 require 'time'
 require 'date'
+require 'numerizer'
+
+require 'chronic/version'
 
 require 'chronic/parser'
+require 'chronic/date'
+require 'chronic/time'
 
 require 'chronic/handler'
 require 'chronic/handlers'
 require 'chronic/mini_date'
-require 'chronic/tag'
 require 'chronic/span'
 require 'chronic/token'
-require 'chronic/grabber'
-require 'chronic/pointer'
-require 'chronic/scalar'
-require 'chronic/ordinal'
-require 'chronic/separator'
-require 'chronic/time_zone'
-require 'chronic/numerizer'
 require 'chronic/season'
 
-require 'chronic/repeater'
+require 'chronic/tag'
+require 'chronic/tags/grabber'
+require 'chronic/tags/ordinal'
+require 'chronic/tags/pointer'
+require 'chronic/tags/scalar'
+require 'chronic/tags/separator'
+require 'chronic/tags/sign'
+require 'chronic/tags/time_zone'
+
+require 'chronic/tags/repeater'
 require 'chronic/repeaters/repeater_year'
 require 'chronic/repeaters/repeater_season'
 require 'chronic/repeaters/repeater_season_name'
@@ -50,7 +56,6 @@ require 'chronic/repeaters/repeater_time'
 #   Chronic.parse('monday', :context => :past)
 #     #=> Mon Aug 21 12:00:00 PDT 2006
 module Chronic
-  VERSION = "0.9.1"
 
   class << self
     # Returns true when debug mode is enabled.
@@ -87,7 +92,7 @@ module Chronic
   end
 
   self.debug = false
-  self.time_class = Time
+  self.time_class = ::Time
   self.locale = :en
 
   require 'chronic/locales/en'
@@ -157,7 +162,7 @@ module Chronic
   # second - Integer second.
   #
   # Returns a new Time object constructed from these params.
-  def self.construct(year, month = 1, day = 1, hour = 0, minute = 0, second = 0)
+  def self.construct(year, month = 1, day = 1, hour = 0, minute = 0, second = 0, offset = nil)
     if second >= 60
       minute += second / 60
       second = second % 60
@@ -175,12 +180,9 @@ module Chronic
 
     # determine if there is a day overflow. this is complicated by our crappy calendar
     # system (non-constant number of days per month)
-    day <= 56 || raise("day must be no more than 56 (makes month resolution easier)")
-    if day > 28
-      # no month ever has fewer than 28 days, so only do this if necessary
-      leap_year_month_days = [31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31]
-      common_year_month_days = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31]
-      days_this_month = Date.leap?(year) ? leap_year_month_days[month - 1] : common_year_month_days[month - 1]
+    day <= 56 || raise('day must be no more than 56 (makes month resolution easier)')
+    if day > 28 # no month ever has fewer than 28 days, so only do this if necessary
+      days_this_month = ::Date.leap?(year) ? Date::MONTH_DAYS_LEAP[month] : Date::MONTH_DAYS[month]
       if day > days_this_month
         month += day / days_this_month
         day = day % days_this_month
@@ -196,7 +198,13 @@ module Chronic
         month = month % 12
       end
     end
-
-    Chronic.time_class.local(year, month, day, hour, minute, second)
+    if Chronic.time_class.name == 'Date'
+      Chronic.time_class.new(year, month, day)
+    elsif not Chronic.time_class.respond_to?(:new) or (RUBY_VERSION.to_f < 1.9 and Chronic.time_class.name == 'Time')
+      Chronic.time_class.local(year, month, day, hour, minute, second)
+    else
+      offset = Time::normalize_offset(offset) if Chronic.time_class.name == 'DateTime'
+      Chronic.time_class.new(year, month, day, hour, minute, second, offset)
+    end
   end
 end