+++ lib/chronic/repeaters/repeater_minute.rb
@@ -2,14 +2,15 @@ module Chronic
   class RepeaterMinute < Repeater #:nodoc:
     MINUTE_SECONDS = 60
 
-    def initialize(type)
+    def initialize(type, options = {})
       super
+      @current_minute_start = nil
     end
 
     def next(pointer = :future)
       super
 
-      if !@current_minute_start
+      unless @current_minute_start
         case pointer
         when :future
           @current_minute_start = Chronic.construct(@now.year, @now.month, @now.day, @now.hour, @now.min + 1)